+++ src/compiler/llvm_codegen_internal_impl.h
@@ -357,6 +357,7 @@ INLINE void llvm_emit_statement_chain(GenContext *c, AstId current)
 {
 	while (current)
 	{
+
 		llvm_emit_stmt(c, ast_next(&current));
 	}
 }